Smart-Medical-Report-Assisstant/requirements.txt
streamlit
PyPDF2
langchain
langchain_google_genai
faiss-cpu
langchain_community
python-dotenv
pypdf
google-generativeai
google-cloud-aiplatform>=1.38
